requirements.txt
anyio==4.3.0
certifi==2024.2.2
deep-translator
googletrans==2.4.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==5.2.0
idna==2.10
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
python-telegram-bot==21.5
requests==2.31.0
sniffio==1.3.1
telepot==12.7
